requirements-base.txt
Django<3.2
django-extensions
django-filter
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gevent
gunicorn
natsort
psutil
pymediainfo
pyyaml
requests
zmq
